[MRESOLVER-153] Make TrackingFileManager use NamedLocks
Step1: drop all SyncContextFactory implementations and indirections, redo the same
functionality using NamedLocks.
